test `go`
